+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/actions/CCv2EnvironmentActions.kt
@@ -20,10 +20,19 @@ package com.intellij.idea.plugin.hybris.tools.ccv2.actions
 
 import com.intellij.idea.plugin.hybris.common.utils.HybrisIcons
 import com.intellij.idea.plugin.hybris.settings.CCv2Settings
+import com.intellij.idea.plugin.hybris.settings.CCv2Subscription
 import com.intellij.idea.plugin.hybris.tools.ccv2.CCv2Service
 import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2Environment
 import com.intellij.idea.plugin.hybris.tools.ccv2.dto.CCv2EnvironmentStatus
+import com.intellij.idea.plugin.hybris.toolwindow.HybrisToolWindowFactory
 import com.intellij.idea.plugin.hybris.toolwindow.ccv2.CCv2Tab
+import com.intellij.idea.plugin.hybris.toolwindow.ccv2.views.CCv2EnvironmentDetailsView
+import com.intellij.openapi.actionSystem.ActionUpdateThread
+import com.intellij.openapi.actionSystem.AnActionEvent
+import com.intellij.openapi.project.DumbAwareAction
+import com.intellij.openapi.util.Disposer
+import com.intellij.openapi.wm.ToolWindow
+import com.intellij.openapi.wm.ToolWindowManager
 
 class CCv2FetchEnvironmentsAction : AbstractCCv2FetchAction<CCv2Environment>(
     tab = CCv2Tab.ENVIRONMENTS,
@@ -33,6 +42,39 @@ class CCv2FetchEnvironmentsAction : AbstractCCv2FetchAction<CCv2Environment>(
         CCv2Service.getInstance(project).fetchEnvironments(subscriptions, onStartCallback, onCompleteCallback)
     }
 )
+
+class CCv2ShowEnvironmentDetailsAction(
+    private val subscription: CCv2Subscription,
+    private val environment: CCv2Environment
+) : DumbAwareAction("Show Environment Details", null, HybrisIcons.CCV2_ENVIRONMENT_DETAILS) {
+
+    override fun getActionUpdateThread() = ActionUpdateThread.EDT
+
+    override fun actionPerformed(e: AnActionEvent) {
+        val project = e.project ?: return
+        val toolWindow = ToolWindowManager.getInstance(project)
+            .getToolWindow(HybrisToolWindowFactory.ID) ?: return
+        val contentManager = toolWindow.contentManager
+        val panel = CCv2EnvironmentDetailsView(subscription, environment)
+        val content = contentManager.factory
+            .createContent(panel, environment.name, true)
+            .also {
+                it.isCloseable = true
+                it.isPinnable = true
+                it.icon = HybrisIcons.CCV2_ENVIRONMENT_DETAILS
+                it.putUserData(ToolWindow.SHOW_CONTENT_ICON, true)
+            }
+
+        Disposer.register(toolWindow.disposable, panel)
+
+        contentManager.addContent(content)
+        contentManager.setSelectedContent(content)
+    }
+
+    override fun update(e: AnActionEvent) {
+        e.presentation.isEnabled = environment.dynatraceLink != null
+    }
+}
 
 abstract class AbstractCCv2ShowEnvironmentWithStatusAction(status: CCv2EnvironmentStatus) : AbstractCCv2ShowWithStatusAction<CCv2EnvironmentStatus>(
     CCv2Tab.ENVIRONMENTS,